bloc for notes implemented
