+++ traces/lib/screens/notes/__notes.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:traces/Models/noteModel.dart';
+import 'package:traces/Models/__noteModel.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:traces/constants.dart';
 import 'package:traces/screens/notes/noteRepository.dart';
