+++ traces/lib/screens/notes/details_bloc/details_bloc.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:meta/meta.dart';
 import 'package:traces/screens/notes/note.dart';
-import 'package:traces/screens/notes/repo/note_repository.dart';
+import 'package:traces/screens/notes/repository/note_repository.dart';
 import './bloc.dart';
 
 class DetailsBloc extends Bloc<DetailsEvent, DetailsState> {
@@ -23,14 +23,15 @@ class DetailsBloc extends Bloc<DetailsEvent, DetailsState> {
       yield* _mapGetNoteDetailsToState(event);
     }else if(event is NewNoteMode){
       yield* _mapNewNoteModeToState(event);
-    }else if(state is EditMode){
+    }else if(event is EditMode){
       yield* _mapEditModeToState(event);
+    }else if(event is SaveNote){
+      yield* _mapSaveNoteToState(event);
     }
   }
 
   Stream<DetailsState> _mapGetNoteDetailsToState(GetNoteDetails event) async*{
     Note note = await _notesRepository.getNoteById(event.noteId);
-
     yield ViewDetailsState(note);
   }
 
@@ -40,5 +41,16 @@ class DetailsBloc extends Bloc<DetailsEvent, DetailsState> {
 
   Stream<DetailsState> _mapEditModeToState(EditMode event) async*{
     yield EditDetailsState(event.note);
+  }
+
+  Stream<DetailsState> _mapSaveNoteToState(SaveNote event) async*{
+    Note note;
+
+    if(event.note.id != null){
+      note = await _notesRepository.updateNote(event.note);
+    }else{
+      note = await _notesRepository.addNewNote(event.note);
+    }
+    yield ViewDetailsState(note);
   }
 }
