+++ traces/lib/screens/notes/note_page.dart
@@ -3,10 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:traces/constants.dart';
 import 'package:traces/screens/notes/notes_bloc/bloc.dart';
-import 'package:traces/screens/notes/note-filter-button.dart';
-import 'package:traces/screens/notes/noteRepository.dart';
-import 'package:traces/screens/notes/notes-view.dart';
-import 'package:traces/screens/notes/repo/firebase_notes_repository.dart';
+import 'package:traces/screens/notes/note_filter_button.dart';
+import 'package:traces/screens/notes/notes_view.dart';
+import 'package:traces/screens/notes/repository/firebase_notes_repository.dart';
 import 'package:traces/colorsPalette.dart';
 import 'package:google_fonts/google_fonts.dart';
 
@@ -16,7 +15,9 @@ class NotesPage extends StatelessWidget{
   @override
   Widget build(BuildContext context) {
     return BlocProvider<NotesBloc>(
-      create: (context) => NotesBloc(notesRepository: FirebaseNotesRepository()),
+      create: (context) => NotesBloc(
+          notesRepository: FirebaseNotesRepository()
+      )..add(GetNotes()),
       child: Scaffold(
         appBar: AppBar(
           title: Text('Notes', style: GoogleFonts.quicksand(textStyle: TextStyle(color: ColorsPalette.grayLight, fontSize: 40.0))),
